refactor:#46 계획 띄워주는 박스 파일 분리
